requirements.txt
streamlit
pandas
matplotlib
openpyxl
plotly
st-gsheets-connection
yfinance
google-api-python-client
google-auth
openpyxl
